backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
websockets>=12.0
librosa==0.10.2.post1
soundfile==0.12.1
numpy>=1.24,<2.0
torch>=2.0
torchaudio>=2.0
demucs>=4.0
openai-whisper>=20231117
music2emo
opencv-python-headless==4.10.0.84
Pillow>=10.0
scipy>=1.11
sentence-transformers>=3.0
moviepy==2.1.1
pydantic>=2.0
pydantic-settings>=2.0
aiofiles==24.1.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
